Split InputParams from RunParams
The intention is to allow some input params, such as problem size,
to have multiple values. Then a RunParam will be created for each
value.
This will make it easier to run scaling and parameter studies.
